Added ChannelsSuite
Issue #119
Includes creating a channel with a user, messaging the channel, attempting to create a duplicate channel, and deleting a channel (admin-only).
Creating different types of channels has not been implemented yet. 
Searching channel using three different ways is still in progress.
Will update the name of file to be more relevant and descriptive.